requirements.txt
aniso8601==9.0.1
blinker==1.6.2
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-WTF==1.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mongoengine==0.27.0
PyJWT==2.7.0
pymongo==4.3.3
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
Werkzeug==2.3.4
WTForms==3.0.1
